Prevent frontend from spuriously starting
The update-keys script (which is called from a cronjob) will
execute an httpd graceful to reload the keys. This is cool
except for when you actually wanted the frontend to stay off
and httpd graceful will cause httpd to start :/

This patch simply makes sure that httpd is running (via
checking httpd status) before executing the graceful. If the
server isn't started, the graceful doesn't run