Updated DQM GUI frontend redirection rules
Removed CAF.
Got rid of really ancient rules for ecal and lumi.
Added online-test to the list of servers we want available from cern
network.
Added two new servers: offline-test and relval-test that should be
available on the test machine (vocms0131)

Rules for testbed are same as for production, with the distinction of
the redirection for production offline and relval, which still give us a
way of accessing the old SLC5 servers.