Removed unused quota file for the online.
From what I can deduct there was once, a long long time ago, some sort of
a Deletion Daemon running on the online. But I've never seen this
working and at least the last year and during Atanas' time, cleaning
happened manually.
This quota file is definitely not in use, so I'm removing it.