Added the Castor backup back to the testing server
This was a much more complicated exercise than originally estimated,
because we nowadays have 3 servers running on the testing/dev machine.
Seeking a balance between functionality vs. simplicity, it will only be
the dev flavor that will run the Castor chain. Since forcing the backup
location of offline-test and relval-test to something different could
become very confusing to people.