Updated the rules for alternate user.
These rules were becoming overly complex and impossible to read.
Because of the removal of the caf, it had now become impossible to even
start a local offline server.
I greatly simplified this to do what I think it is intended to do.
Tested all possible scenario's and I think it should be okay.