Added funct. to remove pickle files on agents stop
In the past we had agents that kept their "state" in memory between
their different wake-up times (like every 15 minutes). Now we changed
two of these agents to be woken up by acron(tab) every X minutes.
Inbetween they just write their entire state (that used to be in memory)
to a pickle file. This is nice, but it means that the state goes on
forever. Even after an update of the daemon or a restart of the server,
the state is preserved.
In the past each month at least the daemons were restarted and their
memory was reset.
We want the same thing to happen for the daemons that write their state
to a file. So now when you stop/start/restart the daemons, the pickle
files are removed and recreated.