Re-arranging workspaces and adding info workspaces (#417)
For both the online and the offline GUI the workspaces are re-ordered
and made more logical and consistent.
We also added layouts for the "Info" plots, so that we could add
documentation. These layouts were then added as workspaces as well, so
that from now on, you can double-click on the HV plot in the summary
page to get more general info on the run, also allowing you to zoom in
to this plot.