Force IPv4 for dev DQMGUI server (#434)
The underlying network layer of DQMGUI is force to use IPv4
network addressing space since the DQMCollector application
is using such version. Usage of the generic 'localhost'
interface could result in IPv6 usage, depending on how the
hosting machine is configured, the facto preventing the Live
source to work properly.