Updated the deploy and manage scripts for CC7
The manage script basically enables you to start the servers and agents
on the new servers.
The deploy script was _not_ changed, to the extend that the acron jobs
that do the backups to CASTOR will not be installed on the new servers.
But of course this will be needed when we do the real migration.
Also some changes to the monitoring scripts. This is basically for the
logging and I just added the new servers where the old servers where.
Also removed some very old servers from those files.